requirements/base.txt
celery[all]==5.3.1
mako==1.2.4
plumbum==1.8.2
boto3==1.28.15
pydantic==2.1.1
pyyaml==6.0.1
cryptography==41.0.2
azure-storage-blob==12.17.0
tap.py==3.1
librabbitmq==2.0.0
requests==2.31.0
filesplit==3.0.2,<4.0
pulpcore-client==3.29.5
pyone==6.6.3
asyncssh==2.13.2
